requirements.txt
instaloader
selenium
pyyaml